==> OG_notebooks/1_merging_notebook.ipynb <==
import pandas as pd

# Merging DataFrames Notebook

For the career village recommender system, we were given 15 CSV's to work with. To make future processing easier and to find significant relationships, we're merging relevant DataFrames. This notebook is specifically dedicated to merging DataFrames. After opening and playing around with the dataframes I realized that merging them wouldn't be as simple as I had hoped. The dataframes don't line up well, and when merging I'd either lose or make up information. In this notebook, I did an initial run at merging the notebooks so that I could delve further into the data, do some cleaning and EDA. After cleaning and EDA, I tweaked how I merged the data for optimal information. For best results, it'd be beneficial to spend more time fitting the datasets together with minimal loss or made-up information. 

For general cleaning, EDA, and modeling I merged:
* answers.csv
* answer_scores.csv
* question_scores.csv
* questions.csv
* tag_questions.csv
* tags.csv
* professionals.csv

## <span style = 'color:red'> include image
    
To do this I did, I first merged the answers CSV's together, the just the questions CSVs, then just the tag CSVs. Once I had three merged notebooks then I merged those notebooks. 

## <span style = 'color:red'> explain how


To further explore the relationships tags have on how questions being answered I merged:
(an to avoid issues of making up or losing information)
## <span style = 'color:red'> List
    
With the data provided, there was only a sparse data dictionary with just a note about what the notebook was. There wasn't an explicit explanation for what the specific columns meant and what was compatible with what. After playing around with a few of the notebooks I assumed that **these** columns were compatible with each other and merged on those columns. 

#### CSV's Used:

# Answers
answers_df = pd.read_csv('./Datasets/provided_csvs/answers.csv')
answer_score_df = pd.read_csv('./Datasets/provided_csvs/answer_scores.csv')

# Questions
q_score_df = pd.read_csv('./Datasets/provided_csvs/question_scores.csv')
ques_df = pd.read_csv('./Datasets/provided_csvs/questions.csv')

# Tags
tag_q_df = pd.read_csv('./Datasets/provided_csvs/tag_questions.csv')
tags_df = pd.read_csv('./Datasets/provided_csvs/tags.csv')

#Professional
pros_df = pd.read_csv('./Datasets/provided_csvs/professionals.csv')

#### Unused CSV's

Given limited time, I only worked with 7/15 notebooks. I didn't include the tag_users.csv because it wasn't compatible with how I merged the other columns and would have been making up information that wasn't necessarily true. The rest of the dataframes weren't used because they weren't important for the specific problem I was seeking to solve and because merging them would skew the information.

#Unused Tags
tag_users_df = pd.read_csv('./Datasets/provided_csvs/tag_users.csv')

# Others
coms_df = pd.read_csv('./Datasets/provided_csvs/comments.csv')
emails_df = pd.read_csv('./Datasets/provided_csvs/emails.csv')
group_mem_df = pd.read_csv('./Datasets/provided_csvs/group_memberships.csv')
groups_df = pd.read_csv('./Datasets/provided_csvs/groups.csv')
school_mem_df = pd.read_csv('./Datasets/provided_csvs/school_memberships.csv')
students_df = pd.read_csv('./Datasets/provided_csvs/students.csv')
matches_df = pd.read_csv('./Datasets/provided_csvs/matches.csv')

#### Throughout this notebook you'll notice I do the same thing for each merge
1. View the notebooks first few rows
2. Check the shape of the dataframe (to ensure it merged the way expected)
3. Rename columns (so that when merging the column names match). *Again, I had to assume these columns were compatible based on examining the data.*
4. Rename columns if the column name will be confusing once merged (i.e. there' a 'score' column in both the questions and answers df)

4. Merging, all the merges are merged on a column shared by both dataframes with a left merge 

* At the end of the notebook I export it as a CSV to the Datasets file

<span style ='color:red'> pic of left merge

## Merging Dataframes with Answers Information

#### `answers_df`

print("shape:", answers_df.shape)
answers_df.head(2)

#### `answer_scores_df`

print("shape:", answer_score_df.shape)
answer_score_df.head()

#### Renaming `id` to `answer_id` to match the answers dataframe for merging

answer_score_df.rename(columns={'id': 'answers_id'}, inplace=True)
answer_score_df.head(1)

#### Merging `answers_df` and `answer_scores`  dataframes on `answer_id`, how ='left', and saving new dataframe as `df_answer`

# Merge two Dataframes on single column 'answers_id'
df_answer = answers_df.merge(answer_score_df, on='answers_id', how='left')
print("shape:", df_answer.shape)
df_answer.head(2)

#### Renaming `score` to `answer_score` to avoid confusion in later merges

df_answer.rename(columns = {'score':'answers_score'}, inplace=True)

---
# Merging DataFrames with Question Information:

#### `q_score_df`

print("shape:", q_score_df.shape)
q_score_df.head(2)

#### Renaming the `id` column in `q_score_df` to match the `questions_id` column in `ques_df` so we can merge the two DataFrames on `questions_id`

q_score_df.rename(columns={'id':'questions_id'}, inplace= True)

#### `ques_df`

print("shape:", ques_df.shape)
ques_df.head(2)

#### Creating a new DataFrame `questions_df` from the merged `ques_df` and `q_score_df`

df_questions = ques_df.merge(q_score_df, on = 'questions_id', how='left')
print("shape:", df_questions.shape)
df_questions.head(2)

#### Renaming `score` to `questions_score` to avoid confusion when merging later

df_questions.rename(columns = {'score':'questions_score'}, inplace=True)

---
---
---

## Merging the Tag Dataframes

### <span style = 'color:red'> if I decide that I'll only use 2/3 tag csv's make sure to note why

Merging the `tag_q_df` and the `tags_df`. We're merging on the `tag_id` with a left merge. We're doing this to keep all the information in the tag_q_df while matching up (and sometimes duplicating) information in the tags_df. 

We're not merging the `tag_users` dataframe because 1) the information isn't crucial to later EDA and modeling and more importantly 2) because we'd either be making up information, duplicating rows that shouldn't be duplicated, or losing significant amounts of information. Since it doesn't add that much value, we're choosing to not merge the dataframe and leave it as a standalone file. 

#### <span style = 'color:red'> insert image is we can

#### Renaming the `tag_questions_question_id` column in `tag_q_df` to match the `questions_id` column in `questions_df` so we can merge the two DataFrames on `questions_id`

print("shape:", tag_q_df.shape)
tag_q_df.head(2)

print("shape:", tags_df.shape)
tags_df.head(2)

#### Renaming `tag_questions_tag_id` to `tag_id` in `tag_q_df` and `tags_tag_id` in `tags_df` so they match and we can merge on them. 

tag_q_df.rename(columns ={'tag_questions_tag_id': 'tag_id'}, inplace = True)
tags_df.rename(columns={"tags_tag_id": "tag_id",  'tags_tag_name':'tag_name'}, inplace=True)

#### Merging the tag columns on `tag_id`. We doing a left merge because we want to keep all the information in the `tag_q_df` and don't mind if the `tags_df` is duplicated or deleted.

df_tags = tag_q_df.merge(tags_df, on = "tag_id", how="left")

print("shape:", df_tags.shape)
df_tags.head(2)

## <span style = 'color:red'> Explain

#### Exporting the tags dataframe so we can do specific EDA on the tags without messing up the other dataframes or merged data

df_tags.to_csv('./Datasets/merged_tag_df.csv',index=False)

#### Renaming the `tag_questions_question_id` column to `questiosn_id` so that it's compatible to the quetions 

df_tags.rename(columns={'tag_questions_question_id': "questions_id"}, inplace = True)

---
---
---

### Merging Tags to Questions Dataframes

# tag_ques_df = df_questions.merge(df_tags, on = 'questions_id', how='left')

# print("shape:", tag_ques_df.shape)
# tag_ques_df.head(2)

#### Renaming `score` to `questions_score` to avoid confusion when merging `tag_ques_df` and answers DataFrames

# tag_ques_df.rename(columns = {'score':'questions_score'}, inplace=True)

# tag_ques_df.shape
# tag_ques_df.head(2)

## Merging `tag_ques_df` to the `df_answers` dataframe and saving as merged

#### Renaming `answers_question_id` to `question_id`  so we can merge the questions and answers `DataFrames`

df_answer.rename(columns = {'answers_question_id' : 'questions_id'}, inplace=True)

#### Merged questions and asnwers DataFrames

print("shape:", df_answer.shape)
df_answer.head(2)

## <span style = 'color:red'> test: We're making it a left merge on the questions df

# merged_df = df_answer.merge(tag_ques_df, on='questions_id', how='left', indicator=True)
# merged_df.head(2)

# merged_df = tag_ques_df.merge(df_answer, on='questions_id', how='left')
# print("shape:", merged_df.shape)
# merged_df.head(2)

qa_merged_df = df_questions.merge(df_answer, on='questions_id', how='left')
print("shape:", qa_merged_df.shape)
qa_merged_df.head(2)

qa_merged_df.to_csv("./Datasets/qa_merged_df.csv", index=False)

## <span style = 'color:red'> This size isn't right, it should be about 55k. For some reasons it's adding both the questions and asnwers data frames. 

#### Exporting as merged csv:

# merged_df.to_csv("./Datasets/merged_df.csv", index=False)

---
---
---
After doing some EDA, I realized that a good dataframe to EDA on would be `answers_df`, `ques_df`, and `tag_q_df`

answers_df.head(2)
answers_df.rename(columns={'answers_question_id' :'questions_id'}, inplace=True)

ques_df.head(2)

qa_merge = ques_df.merge(answers_df, on = 'questions_id', how = 'left')

qa_merge.head(2)

qa_merge.isnull().sum()

qa_merge['was_answered'] = qa_merge['answers_id'].notnull().astype(int)

qa_merge = qa_merge.drop(columns=['answers_id', 'answers_author_id','answers_date_added', 'answers_body'])

print("shape:", qa_merge.shape)
qa_merge.head(2)

tag_q_df.head()

tag_q_df['has_tag'] = tag_q_df['tag_questions_question_id'].notnull().astype(int)
tag_q_df.head(2)

has_tag_df = tag_q_df.drop(columns = 'tag_id')
has_tag_df = has_tag_df.drop_duplicates()

print("shape:", has_tag_df.shape)
has_tag_df.head(2)

has_tag_df.rename(columns = {'tag_questions_question_id': 'questions_id'}, inplace=True)
has_tag_df.head(2)

merged_qa_tag_df = qa_merge.merge(has_tag_df, on="questions_id", how='left')
print("shape:", merged_qa_tag_df.shape)
merged_qa_tag_df.head(2)

#### Exporting `merged_qa_tag_df` to csv for EDA

# merged_qa_tag_df.to_csv('./Datasets/merged_qa_tag_df.csv', index=False)

---
---
---

df_answer.head(2)

df_answer['was_answered'] = 1

was_ans_df = df_answer[['questions_id', 'was_answered', 'answers_score']]

ques_df.head(2)

q_answered_df = df_questions.merge(was_ans_df ,on='questions_id', how='left')
q_answered_df.head()

q_answered_df.head()

q_answered_df.to_csv('./Datasets/merged_4_modeling.csv', index=False)
