==> OG_notebooks/2_cleaning_4_modeling_notebook.ipynb <==
# Cleaning Notebook

Before conducting EDA and building the data needs to be formatted in a way that won't throw errors. In a combination of intial exploration and later EDA the following steps are needed:

1. Creating Columns 
2. Converting to Appropriate Datatype
3. Handeling Nulls
4. Cleaning Text
5. Dealing Duplicate Values Left Over from Merging Madness

### Importing libarires

For this notebook we'll use pandas, regex (for cleaning text), and seaborn (for visualizations).

import pandas as pd
import regex as re
import seaborn as sns
import matplotlib.pyplot as plt

# from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
# from nltk.tokenize import RegexpTokenizer
# from nltk.sentiment.vader import SentimentIntensityAnalyzer
# from sklearn.feature_extraction import stop_words, text

# This creates HD resolution for visualizations

%config InlineBackend.figure_format = 'retina'

#### Reading in the data, checking the shape and first few rows.

An initial look at the dataframe reveals that the text for `answers_body` has several HTML artifacts. In the NLP cleaning later we'll remove these as well as clean other text columns are a precaution. 

<span style ='color:red'> include? Or move to merge notebook?? Should start off with a bang
One of the things we notice off the bat is that size is 180K rows. From the merging we notebooks we know that this is an error from merging. Through EDA and cleaning this error persists and essentially gives us unbalanced classes while merging. Due to limited time, we weren't able to fix this error. For better results, I recommend fixing this issue. 

data = pd.read_csv('./Datasets/merged_4_modeling.csv')
print("shape:", data.shape)
data.head(2)

Data.describe() reveals a quick overview of the numeric columns. A peruse over the stats shows that there are no significant outliers that need to be cleaned.

data.describe()

By checking the datatypes I discovered:
1. Dates are saved as *objects* we'll change these laters
2. Id's are stored as *objects*, in checking the dataframe above we see that id's use both numbers and letters and shouldn't be converted to another datatype. That said, when doing our NLP training these should be treated differently than the other *object* columns

data.dtypes

#### Creating a `date_cols` list so we can loop through and convert dates to date_time data type

date_cols = []

for cols in data.columns:
    if "date" in cols:
        date_cols.append(cols)
        
date_cols

#### Converting Date Columns to *Date* type

for cols in date_cols:
    data[cols] = pd.to_datetime(data[cols])
    
data.dtypes

#### Extracting columns which are *object* oriented type and saving them as a list called `str_cols`. We do this so we can easily process all the *object* types in the `cleaning_text` function.


str_cols = data.select_dtypes(include ='object').columns
str_cols

After looking at the various *object* type columns I realize that the `id` columns shouldn't be treated as a string. Since they include both numbers and letters, we can't convert them into a interger. So instead let's create a list, `text_cols` that has all *object* type columns excluding the `id` columns.

#### Using a for loop to create `text_cols` list which is a list of columns that are text based (*object* columns that aren't `id` columns)


text_cols = [] #create a list

for cols in str_cols: # looping through the `str_cols` variable

    if "id" not in cols: # if `id` isn't in the name 
        text_cols.append(cols) # append to `text_cols` list

text_cols

### Handling Nulls:

Many of our models and code doesn't work when null values are present. We need to either delete, or fill the nulls so that later we won't incur any errors.

A major contributor for nulls in this dataset is from how the data was given to us. There were 15 datasets that didn't perfectly match. We had to choose to either duplicate/create information or to loose information. Since making up information creates more issues, we choose to deal with the large number of nulls values. 

First we need to check for nulls.

## <span style = 'color:red'> For now we're just filling nulls.

data.isnull().sum()

#### Filling nulls

When merging, all the `was_answered` were '1' representing that it was answered, with no nulls. Since we did a left merge on the questions dataset we know all the nulls are where questions weren't answered. Therefor I'm filling `was_answered` nulls with 0.

data['was_answered'].fillna(0, inplace=True)

## Natural Language Processing Cleaning

Since we want to explore the text data and build models form the text we need to clean the text to get the best results. 

#### <span style='color:red'> explain more

#### Setting up a Cleaning Text Function: 

The function takes in a column and cleans the text for preproccessing. It removes html artifacts as well as punctuation and numbers, and converts the text to all lower case. 

def cleaning_text(df, df_col):
    """
    df: is the name of Dataframe 
    df_col: takes in a column name formated as string i.e. "column_name" 
    
    This function takes a column and cleans the text for that column. 
    It removes HTML artifacts suchs as <p> and <br> as as well as punctuation, 
    and numbers preparing it for processing and modeling. 
    In addition it makes all the text lower case.
    It utilizes the .replace method as well as regex. 
    It outputs the top 2 rows
    
    If you haven't yet, the code to import regex is:
    import regex as re 
    #you need to regex import as re or else this code will break
    """

    # Uses .str.reaplce
    df[df_col] = df[df_col].str.replace("<p>", "") #removes <p>
    df[df_col] = df[df_col].str.replace("</p>","") #removes </p>
    df[df_col] = df[df_col].str.replace("<br>", "") #removes <br>
    df[df_col] = df[df_col].str.replace("\n", "")   #removes <\n>
    
    #Makes everything lower case
    df[df_col] = df[df_col].str.lower()
    
    # Using regex and lambda 
    df[df_col] = df[df_col].map(lambda x: re.sub('\/\/', ' ', x)) # Removing line breaks
    df[df_col] = df[df_col].map(lambda x: re.sub('[\\][\']', '', x)) # Removing apostrophes
    df[df_col] = df[df_col].map(lambda x: re.sub('[^\w\s]', ' ', x)) # Removing all punctuation 
    df[df_col] = df[df_col].map(lambda x: re.sub('\xa0', ' ', x)) # removing xa0
    df[df_col] = df[df_col].map(lambda x: re.sub('\s[\/]?r\/[^\s]+', ' ', x)) # removing mentions of any subreddit
    df[df_col] = df[df_col].map(lambda x: re.sub('http[s]?:\/\/[^\s]*', ' ', x)) # removing urls
    
    # Keeping numbers for now, if we want to strip numbers, use the below
    df[df_col] = df[df_col].map(lambda x: re.sub("[^a-zA-Z]", " ", x)) # Removes all numbers only keeping letters
    
    #Displays the top 2 rows
    return df[df_col].head(2)

#### For loop interates through `text_cols` list (all the *object* oriented columns) and cleans the text using `cleaning_text()` function. 

for cols in text_cols:
    if data[cols].isnull().sum() == 0: #If there aren't any nulls 
        cleaning_text(data, cols) #Call `cleaning_text` function
    else:
        print(f"{cols} has null values, so we're filling with 'none', then calling `cleaning_tex()` function") #print which columns have nulls 
        
        #Filling nulls with 'none'
        data[cols].fillna('none', inplace =True)
        
        #Then calling the function
        cleaning_text(data, cols) #Call `cleaning_text` function

data.head(2)

## Exporting Data

data.to_csv('./Datasets/cleaned_4_modeling.csv', index=False)
